+++ src/components/ProductsSlider/ProductsSlider.module.scss
