--- src/components/SliderProducts/SliderProducts.scss
@@ -1,1 +1,0 @@
-
